feat: basic snapshot encoding

+++ crates/loro-internal/src/container/list/list_container.rs
@@ -141,7 +141,7 @@ impl ListContainer {
             let id = store.next_id();
             let op = Op::new(
                 id,
-                InnerContent::List(InnerListOp::new_del(pos, len)),
+                InnerContent::List(InnerListOp::new_del(pos, len as isize)),
                 self.idx,
             );
             store.append_local_ops(&[op]);
